Create and rename form component
